% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{diversity}
\alias{diversity}
\title{Shannon diversity index}
\usage{
diversity(x, two.power = FALSE)
}
\arguments{
\item{x}{Vector of memberships}

\item{two.power}{Logical, whether return to 2 to the power of Shannon diversity index.}
}
\value{
A numeric value of Shannon diversity index \eqn{\mathrm{E}} or \eqn{2^{\mathrm{E}}}.
}
\description{
Calculate the Shannon diversity index of the memberships of each observation. The base of the logarithm is 2.
}
\details{
Given the membership vector of the \eqn{i} observation \eqn{h_i}, the entropy is defined as
\deqn{\mathrm{E}(h_{i \cdot}) = -\sum_{r=1}^k h_{ir} \cdot log_b h_{ir},}
where b is the base of the logarithm used.

Specifically, in the case of \eqn{h_{ir}=0}, the value of \eqn{0 \mathrm{log}_2 (0)} is taken to be 0 based on the limit theory.

Taken 2 to the power of \eqn{\mathrm{E}(h_{i \cdot})}, \eqn{1 \leq 2^{\mathrm{E}(h_{i \cdot})} \leq k},
the maximum is achieved when \eqn{h_{i1} = \dots = h_{ik} = \frac{1}{k}}.
}
\examples{
# Memberships vector
membership1 <- c(0.1, 0.2, 0.3, 0.4)
diversity(membership1)
diversity(membership1, two.power = TRUE)

# Memberships matrix
membership2 <- matrix(c(0.1, 0.2, 0.3, 0.4, 0.3, 0.2, 0.4, 0.1, 0.2, 0.3, 0.1, 0.4), nrow=3, ncol=4, byrow=T)

E <- rep(NA, nrow(membership2))
for(i in 1:nrow(membership2)){
  E[i] <- diversity(membership2[i,])
}
E
}
\author{
Wenxuan Liu
}
