% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{fit_boot}
\alias{fit_boot}
\title{Example results of bootstrap.}
\format{
A list of bootstrap result, including the values of estimated prototype matrix (\eqn{W}),
the lower bound of confidence intervals and the upper bound of confidence intervals.
}
\usage{
fit_boot
}
\description{
A list of the results for bootstrap example.
}
\keyword{datasets}
