% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{fit_gap}
\alias{fit_gap}
\title{Example results of gap statistic.}
\format{
A list of gap statistic result, including the gap value vector,
the optimal number of prototypes/clusters and the Standard error vector.
}
\usage{
fit_gap
}
\description{
A list of the results for gap statistic example for \eqn{k=1, 2, ..., 10}.
}
\keyword{datasets}
