% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{fit_SSMF}
\alias{fit_SSMF}
\title{Example results of SSMF.}
\format{
A list with 10 items, each item is a results of SSMF,
containing the values of the estimated prototype matrix (\eqn{W}) and
the estimated membership matrix (\eqn{H}) matrix and the value of
the residuals sum of square (SSE).
}
\usage{
fit_SSMF
}
\description{
A list of the results for SSMF example for \eqn{k=1, 2, ..., 10}.
}
\keyword{datasets}
