% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gap}
\alias{gap}
\title{Gap statistic algorithm.}
\usage{
gap(data, rss, meth = c("kmeans", "uniform", "dirichlet", "nmf"), itr = 50)
}
\arguments{
\item{data}{Data matrix or data frame.}

\item{rss}{Numeric vector, sum of square errors/residual sum of squares at \eqn{1:k} of ssmf model}

\item{meth}{Character, specification of method to initialise the \eqn{W} and \eqn{H} matrix, see 'method' in init().}

\item{itr}{Integer, number of Monte Carlo samples.}
}
\value{
\code{gap} Gap values.

\code{optimal.k} The optimal number of prototypes/clusters that gap statistics suggested.

\code{standard.error} Standard error
}
\description{
Estimating the number of prototypes/clusters in a data set via the gap statistics.
}
\details{
This gap statistics technique is selecting the biggest difference between the original RSS and the RSS under an appropriate null reference distribution of the data, which is defined to be
\deqn{\mathrm{Gap}(k) = \frac{1}{B} \sum_{b} \log(\mathrm{RSS}^*_{kb}) - \log(\mathrm{RSS}_{k})}

where \eqn{B, (b=1, 2, ..., B)} is the number of sampled null reference distributions of the data;
\eqn{RSS^*_{kb}} is the residual sum of squares of the \eqn{b^\mathrm{th}} null reference distributions of the data that fitted in the SSMF model;
\eqn{RSS_{k}} is the residual sum of squares of the original data \eqn{X}. The estimated gap suggests the number of prototypes via

\deqn{\hat{k} = \mathrm{smallest} \ k \  \mathrm{such \ that} \ \mathrm{Gap}(k) \geq \mathrm{Gap}(k+1) - s_{k+1}}

where \eqn{s_{k+1}} is standard error that is defined as

\deqn{s_{k+1}=sd_k \sqrt{1+\frac{1}{B}}}

and \eqn{sd_k} is the standard deviation:

\deqn{sd_k=\{ \frac{1}{B} \sum_{b} [\log(RSS^*_{kb})-\frac{1}{B} \sum_{b} \log(RSS^8_{kb})]^2 \}^\frac{1}{2}}.
}
\examples{
# example code
data <- SimulatedDataset

k <- 6

rss <- rep(NA, k)
for(i in 1:k){
  rss[i] <- ssmf(data = data, k = i)$SSE
}

gap(data = data, rss = rss)
}
\references{
Tibshirani, R., Walther, G., & Hastie, T. (2001). Estimating the Number of Clusters in a Data Set via the Gap Statistic. Journal of the Royal Statistical Society. Series B (Statistical Methodology), 63(2), 411–423. http://www.jstor.org/stable/2680607
}
\author{
Wenxuan Liu
}
