% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{init}
\alias{init}
\title{Initialise the membership matrix \eqn{H} or prototype matrix \eqn{W}.}
\usage{
init(
  data = NULL,
  k = NULL,
  method = c("kmeans", "uniform", "dirichlet", "nmf")
)
}
\arguments{
\item{data}{Data matrix or data frame.}

\item{k}{The number of prototypes/clusters.}

\item{method}{Character: 'kmeans', 'uniform', 'dirichlet' or 'nmf'. If there are more than one methods,
the default is selecting the first method in the vector.}
}
\value{
Initialised \eqn{H}, \eqn{W} matrix.
}
\description{

}
\details{
'kmeans': create \eqn{W} matrix using the centres of the kmeans outputs; create \eqn{H} matrix by converting the classification into a binary matrix.

'uniform': create the \eqn{H} matrix by sampling the values from uniform distribution and making the matrix as unit simplex; group the observations with their maximum memberships
and create the \eqn{W} matrix by combining the mean vector in each group.

'dirichlet': create the \eqn{H} matrix by sampling the values from dirichlet distribution; group the observations with their maximum memberships
and create the \eqn{W} matrix by combining the mean vector in each group.

'nmf': create \eqn{W} matrix using the matrix of basis components from NMF model; create the \eqn{H} matrix using the coefficient matrix from NMF model.
}
\examples{
# example code
init(data = SimulatedDataset, k = 4, method = 'kmeans')
}
\author{
Wenxuan Liu
}
