% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sARI}
\alias{sARI}
\title{Soft adjust Rand index function.}
\usage{
sARI(partition1, partition2)
}
\arguments{
\item{partition1}{Numeric matrix/data frame of the (posterior) probabilities of assignment of observations in partition 1, memberships matrix.}

\item{partition2}{Numeric matrix/data frame of the (posterior) probabilities of assignment of observations in partition 2, memberships matrix.}
}
\value{
Soft adjust rand index.
}
\description{
Soft adjust rand index, a soft agreement measure for class partitions incorporating assignment probabilities
}
\details{
A soft adjusted Rand index (sARI) is proposed to compare the two soft partitions (e.g., the true and estimated fuzzy memberships).
Let \eqn{p_{rci} = h_{ri} \cdot h_{ci}} be the posterior probabilities of assignment of observation \eqn{i} to \eqn{r^{th}}
cluster in the true partition and to \eqn{c^{th}} cluster in the estimated partition, that is \eqn{h_{ri}} and \eqn{h_{ci}} are the given true and estimated fuzzy membership of sample \eqn{i}.
we have the sum of the posterior probabilities of assignment to \eqn{r^{th}} cluster in the true partition and to \eqn{c^{th}} cluster in the estimated partition over all observations.

\deqn{p_{rc.} = \sum_{i=1}^n p_{rci};}

the sum of the posterior probabilities for belonging to assignment to \eqn{r^{th}} cluster in the true partition over all observations.
\deqn{p_{r..} = \sum_{c=1}^C \sum_{i=1}^n p_{rci};}

the sum of the posterior probabilities for belonging to assignment to \eqn{c^{th}} cluster in the estimated partition over all observations.
\deqn{p_{.c.} = \sum_{r=1}^R \sum_{i=1}^n p_{rci}.}

The sARI for true and estimated soft partitions is defined as
\deqn{\mathrm{sARI}(H_{true}, H_{est}) = \frac{\sum_{r,c} \frac{\Gamma(p_{rc.}+1)}{\Gamma(p_{rc.}-1)} - \frac{1}{n(n-1)} \Lambda_{rc}}{\frac{1}{2} \Lambda_{rc} - \frac{1}{n(n-1)} \Lambda_{rc}}}

where \eqn{\Lambda_{rc} = \sum_{r} \frac{\Gamma(p_{r..}+1)}{\Gamma(p_{r..}-1)} + \sum_{c} \frac{\Gamma(p_{.c.}+1)}{\Gamma(p_{.c.}-1)}}. A larger sARI illustrates the more similar between these two partitions.
}
\references{
Flynt, A., Dean, N. & Nugent, R. sARI: a soft agreement measure for class partitions incorporating assignment probabilities. Adv Data Anal Classif 13, 303–323 (2019). \url{https://doi.org/10.1007/s11634-018-0346-x}
}
\author{
Wenxuan Liu
}
