% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{SimulatedDataset}
\alias{SimulatedDataset}
\title{A simulated metabolomic dataset.}
\format{
A data frame with 177 rows and 138 columns.
}
\usage{
data(SimulatedDataset)
}
\description{
A simulated metabolomic data set containing 138 variables for 177 individuals.
}
\keyword{datasets}
