% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{SimulatedMemberships}
\alias{SimulatedMemberships}
\title{A simulated membership matrix}
\format{
A data frame with 177 rows and 4 columns.
}
\usage{
data(SimulatedMemberships)
}
\description{
A simulated membership matrix containing 4 cluster memberships for 177 individuals.
}
\keyword{datasets}
