% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{SimulatedPrototypes}
\alias{SimulatedPrototypes}
\title{A simulated prototype matrix}
\format{
A data frame with 4 rows and 138 columns.
}
\usage{
data(SimulatedPrototypes)
}
\description{
A simulated prototype matrix containing 4 cluster prototypes.
}
\keyword{datasets}
