% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap algorithm function.}
\usage{
bootstrap(data, k, H, mtimes = 50)
}
\arguments{
\item{data}{Data matrix or data frame.}

\item{k}{The number of prototypes/clusters.}

\item{H}{Matrix, input \eqn{H} matrix to start the algorithm. Usually the H matrix here is the outputs of the function ssmf( ).
If H is not input, the bootstrapped \eqn{W} matrix might have different prototypes orders from the outputs of the function ssmf( ).}

\item{mtimes}{Integer, number of bootstrap samples. Default number is 50.}
}
\value{
\code{W.est} The \eqn{W} matrix estimated by bootstrap.

\code{lower} Lower bound of confidence interval.

\code{upper} Upper bound of confidence interval.
}
\description{
Bootstrap resampling approach to estimated the confidence interval.
}
\details{
Create bootstrap samples of size \eqn{n} by sampling from the dataset with replacement and repeat the steps \eqn{M} times.
The \eqn{m^{th}} bootstrap sample is denoted as
\deqn{X^{*(m)}=(x_1^{*m}, x_2^{*m},...,x_n^{*m}),}

where each \eqn{x_i^{*m}} is a random selection from the original observation.

Then, apply the SSMF algorithm to each bootstrap sample and calculate the \eqn{m^{th}} bootstrap replicate of the prototypes matrix,
which is denoted as \eqn{W^{*(m)}}

The estimate standard deviation of \eqn{M} bootstrap replicates can be calculated by

\deqn{sd(W^*) = \sqrt {\frac{1}{M-1} \sum_{m=1}^{M} [W^{*(m)}-\overline{W}^{*}]^2}}

where \eqn{\overline{W}^{*}=\frac{1}{M} \sum_{m=1}^{M} W^{*(m)}}. Therefore, the 95\% CIs for the prototypes can be calculated by

\deqn{(\overline{W}^{*}-t_{(0.025, M-1)} \cdot sd,\ \overline{W}^{*}+t_{(0.975, M-1)} \cdot sd),}
where \eqn{t_{(0.025, n-1)}} and \eqn{t_{(0.975, n-1)}} is the quantile of student \eqn{t} distribution with 95\% significance and \eqn{(M-1)} degree freedom.
}
\examples{
# example code
data <- SimulatedDataset

k <- 4

fit <- ssmf(data = data, k = k)

bootstrap(data = data , k = k, H = fit$H)
}
\references{
STINE, R. (1989). An Introduction to Bootstrap Methods: Examples and Ideas. Sociological Methods & Research, 18(2-3), 243-291. \url{https://doi.org/10.1177/0049124189018002003}
}
\author{
Wenxuan Liu
}
