% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{init}
\alias{init}
\title{init}
\usage{
init(k, n, data = NULL, method = c('dirichlet', 'uniform', 'kmeans', 'nmf'))
}
\arguments{
\item{k}{The number of prototypes/clusters.}

\item{n}{The number of observations.}

\item{data}{Numeric matrix of data.}

\item{method}{character: 'dirichlet', 'uniform', 'kmeans' or 'nmf'.}
}
\value{
Initialised \eqn{H}, \eqn{W} matrix
}
\description{
This function initialise the \eqn{H_{n \times k} = (h_1, \dots, h_i, \dots, h_n)} matrix
and \eqn{W_{k \times p} = (w_1, \dots, w_j, \dots, w_p)^T} matrix to start the SSMF model.
This function is often used in conjunction with the function ssmf().
}
\details{
'dirichlet': sampling \eqn{H} matrix from dirichlet distribution, grouping the observations with their maximum memberships in \eqn{h_i} and calculate the means in each group as prototype vectors of \eqn{W} matrix;

'uniform': sampling \eqn{H} matrix from uniform distribution, \eqn{W} matrix is created in the same way of 'dirichlet';

'kmeans': applying the centres of the kmeans clustering to create \eqn{W} matrix and converts the classification into a \eqn{H} matrix;

'nmf': using the matrix of basis components and the coefficient matrix of an NMF model to create \eqn{W} matrix and \eqn{H} matrix.
}
\author{
Wenxuan Liu
}
