% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{init}
\alias{init}
\title{Initialise the memberships matrix \eqn{H} or prototype matrix \eqn{W}.}
\usage{
init(
  data = NULL,
  k = NULL,
  method = c("kmeans", "uniform", "dirichlet", "nmf")
)
}
\arguments{
\item{data}{Data matrix or dataframe.}

\item{k}{The number of prototypes/clusters.}

\item{method}{Character: 'kmeans', 'uniform', 'dirichlet' or 'nmf'. If there are more than one methods,
the default selection is the first method in the vector.}
}
\value{
Initialised \eqn{H}, \eqn{W} matrix
}
\description{
This function initialise the \eqn{H_{n \times k} = (h_1, \dots, h_i, \dots, h_n)} matrix
and \eqn{W_{k \times p} = (w_1, \dots, w_j, \dots, w_p)^T} matrix to start the SSMF model.
This function is often used in conjunction with the function ssmf(). Also, the codes can be run separately from the function
ssmf(). This function returns to simplex-structured \eqn{H} and prototype \eqn{W} matrix.
}
\details{
'kmeans': applying the centres of the kmeans clustering to create \eqn{W} matrix and converts the classification into a \eqn{H} matrix;

'uniform': sampling \eqn{H} matrix from uniform distribution, grouping the observations with their maximum memberships in \eqn{h_i}
and calculate the means in each group as prototype vectors of \eqn{W} matrix;

'dirichlet': sampling \eqn{H} matrix from dirichlet distribution, grouping the observations with their maximum memberships in \eqn{h_i}
and calculate the means in each group as prototype vectors of \eqn{W} matrix;

'nmf': using the matrix of basis components and the coefficient matrix of an NMF model to create \eqn{W} matrix and \eqn{H} matrix.
}
\examples{
# example code
init(data = SimulatedDataset, k = 4, method = 'kmeans')
}
\author{
Wenxuan Liu
}
