% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sARI}
\alias{sARI}
\title{Soft adjusted Rand index.}
\usage{
sARI(partition1, partition2)
}
\arguments{
\item{partition1}{Numeric matrix/data frame of the probabilities of assignment of observations in partition 1 (membership matrix).}

\item{partition2}{Numeric matrix/data frame of the probabilities of assignment of observations in partition 2 (membership matrix).}
}
\value{
Soft adjusted Rand index.
}
\description{
Soft adjusted Rand index, a soft agreement measure for class partitions incorporating assignment probabilities.
}
\references{
Flynt, A., Dean, N. & Nugent, R. (2019) sARI: a soft agreement measure for class partitions incorporating assignment probabilities. Adv Data Anal Classif 13, 303–323 (2019). <doi:10.1007/s11634-018-0346-x>
}
\author{
Wenxuan Liu
}
